set eventmask for buildstats-summary bbclass
Issue: TA53702

Bitbake bbclass event handlers should have the eventmask set
for the relevant events instead of having all events dispatched
to the handler, then having the handler filter the appropriate
events.  The latter is an inefficiency and not recommended.  Also
due to a problem with bitbake [bugzilla 6332], unmasked event
handling breaks the toaster interface.

This commit applies to the buildstats-summary event handler.

Signed-off-by: Dave Lerner <dave.lerner@windriver.com>