Parse properly if image file location is not an absolute path
Issue: LIN7-1016

If image file location is a relative path, the file
"layers/wrlinux/scripts/create-usb.pl" can not parse it properly.
It can only parse a absolute path.

Now that we have expanded the regexp to include lines which do not
start with a '/' we have to ensure we only look at lines returned
by fdisk which contain partition information

The items of partition information are separated by white space and the
original regular expression can not work well any more.
So the expression to parse partition information is changed to
^\S+\s+\*?\s+(\d+)\s+(\d+)\s+(\d+) from ^\/.*?\s.*?(\d+).*?(\d+).*?(\d+).
The file name may contains some characters out of "\w" and "\S" is better.
The new regular expression is more accurate.

Signed-off-by: Jian Liu <jian.liu@windriver.com>
Signed-off-by: Roy.Li <rongqing.li@windriver.com>