tcf-agent: config-target.pl boots up tcf-agent
Issue: LIN7-1254

This change is to the script that is invoked by 'make start-target'
to start a qemu session for a bsp, config-target.pl.  The two
changes related to the switch from workbench 3x to workbench 4:
1) Remove the UMA=1 kernel option references, the usermode agent is no
longer supported.
2) Add the TCF=1 kernel option reference, to enable tcf-agent for qemu
targets to boot.

The tcf-agent init script that is installed with tcf-agent won't start
tcf-agent unless the command line option includes TCF=1.

Signed-off-by: Dave Lerner <dave.lerner@windriver.com>
Signed-off-by: Jason Wessel <jason.wessel@windriver.com>