distro/wrlinux.conf: Add -fno-optimize-sibling-calls to -mapcs-frame usage
apcs-frames are deprecated in GCC, as such not everything is expected to
work with them.  When VFP or neon is enabled, optimized sibling calls can
cause register corruption.

We still require apcs-frames in order to do some debugging and analysis
activites, so be safe and always disable the siblings calls when apcs
frames are used.

Issue: LIN7-1642 TCWRLGCC-5 LIN7-2730

Signed-off-by: Peter Seebach <peter.seebach@windriver.com>
Signed-off-by: Mark Hatle <mark.hatle@windriver.com>