rpm: force db verssion for native builds
Issue: LIN7-3104

If libdb6-0.so is on the host, rpm-native will find it and try to
use it.  We force configure to assume the version of libdb we want to
use is available.  We need to do this because the version of db we
put on targets must be the same as rpm-native.

Signed-off-by: Joe Slater <jslater@windriver.com>