create-usb.pl: Update to use the MACHINE, if set, to find the right kernel/fs
Issue: LIN8-1401

Use the environment 'MACHINE' to augment the wild card search for the image
and filesystem.  If MACHINE is not set, then previous behavior is retained.

The machine value should be unique enough to avoid incorrect matches.  The
format of this file/link is always <machine>-<type>...  The script is expecting
either <machine> or a wild card, and the type of bzImage.

The filesystem image could match the wrong rootfs in some cases.
This value is only used when the user passes into the script --tarfiles.  The
user may also pass the correct name in with --rootfs=...  And in either case
will be prompted that the automatic or manual selection is correct before it
is used.

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>