config-target.pl: Work with newer versions of sudo which agressively strip env
The BUILDDIR variable must be passed in the environment in order for
the automatic tun/tap device startup to work properly.

Also enforce the limit of tun/tap instances to 50 as that is what the
start-target system supports.  There are other ways to work around
this limitation by creating additional devices and using static
configurations.

[ Issue: LIN8-1989 ]

Signed-off-by: Jason Wessel <jason.wessel@windriver.com>